+++ src/Module/Ship/View/ShowRepairOptions/ShowRepairOptions.php
@@ -59,8 +59,8 @@ final class ShowRepairOptions implements ViewControllerInterface
         $game->setTemplateVar('ERROR', false);
         $game->setTemplateVar('ROUNDS', $this->shipWrapperFactory->wrapShip($ship)->getRepairDuration());
 
-        $game->setTemplateVar('SPARE_PARTS_ONLY', (int)((RepairTaskEnum::SPARE_PARTS_ONLY_MIN + RepairTaskEnum::SPARE_PARTS_ONLY_MAX) / 2));
-        $game->setTemplateVar('SYSTEM_COMPONENTS_ONLY', (int)((RepairTaskEnum::SYSTEM_COMPONENTS_ONLY_MIN + RepairTaskEnum::SYSTEM_COMPONENTS_ONLY_MAX) / 2));
-        $game->setTemplateVar('BOTH', (int)((RepairTaskEnum::BOTH_MIN + RepairTaskEnum::BOTH_MAX) / 2));
+        $game->setTemplateVar('SPARE_PARTS_ONLY', (RepairTaskEnum::SPARE_PARTS_ONLY_MIN + RepairTaskEnum::SPARE_PARTS_ONLY_MAX) / 2);
+        $game->setTemplateVar('SYSTEM_COMPONENTS_ONLY', (RepairTaskEnum::SYSTEM_COMPONENTS_ONLY_MIN + RepairTaskEnum::SYSTEM_COMPONENTS_ONLY_MAX) / 2);
+        $game->setTemplateVar('BOTH', (RepairTaskEnum::BOTH_MIN + RepairTaskEnum::BOTH_MAX) / 2);
     }
 }
